Moved Wizard to own component
